backend/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3  # Add explicit version
psycopg2-binary==2.9.3
python-dotenv==0.19.2
flask-cors==3.0.10  # Add this line
